+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Detection of WordPress theme Divi.
+- Detection of WordPress plugin Classic Editor.
 
 ## [1.0.0]